requirements.txt
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.3
Interface==2.11.1
invoke==2.2.1
librouteros==3.4.1
paramiko==4.0.0
pycparser==2.23
PyNaCl==1.6.1
python-dotenv==1.2.1
RouterOS-api==0.21.0
toml==0.10.2
zope.event==6.1
zope.interface==8.1.1
zope.schema==8.1
